# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9050
  description: Generated server url
paths:
  /api/vendedores/carga:
    post:
      tags:
      - vendedor-controller
      operationId: cargarVendedoresDesdeExcel
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/distribuciones/{procesoId}/simular:
    post:
      tags:
      - distribucion-controller
      operationId: simular
      parameters:
      - name: procesoId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulacionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendedorSimuladoDTO"
  /api/vendedores/{procesoId}:
    get:
      tags:
      - vendedor-controller
      operationId: listarVendedoresValidos
      parameters:
      - name: procesoId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendedorResponseDTO"
  /api/distribuciones/{procesoId}/pdfs:
    get:
      tags:
      - distribucion-controller
      operationId: descargar
      parameters:
      - name: procesoId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
  /api/vendedores:
    delete:
      tags:
      - vendedor-controller
      operationId: eliminarVendedores
      responses:
        "200":
          description: OK
components:
  schemas:
    RangoCortadoDTO:
      type: object
      properties:
        inicio:
          type: integer
          format: int32
        fin:
          type: integer
          format: int32
    SimulacionRequestDTO:
      required:
      - vendedores
      type: object
      properties:
        poolSenete:
          type: array
          items:
            $ref: "#/components/schemas/RangoCortadoDTO"
        poolTelebingo:
          type: array
          items:
            $ref: "#/components/schemas/RangoCortadoDTO"
        vendedores:
          type: array
          items:
            $ref: "#/components/schemas/VendedorInputDTO"
        mezclar:
          type: boolean
        inicioSenete:
          type: integer
          format: int32
        inicioTelebingo:
          type: integer
          format: int32
        fechaSorteoSenete:
          type: string
          format: date
        fechaSorteoTelebingo:
          type: string
          format: date
    VendedorInputDTO:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        cantidadSenete:
          type: integer
          format: int32
        terminacionSenete:
          type: integer
          format: int32
        cantidadTelebingo:
          type: integer
          format: int32
        terminacionTelebingo:
          type: integer
          format: int32
    VendedorSimuladoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        rangosSenete:
          type: array
          items:
            type: string
        rangosTelebingo:
          type: array
          items:
            type: string
    VendedorResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        deuda:
          type: number
        cantidadSenete:
          type: integer
          format: int32
        resultadoSenete:
          type: integer
          format: int32
        cantidadTelebingo:
          type: integer
          format: int32
        resultadoTelebingo:
          type: integer
          format: int32